+++ src/Serializers/Native.php
@@ -3,6 +3,8 @@
 namespace Rubix\Server\Serializers;
 
 use Rubix\Server\Message;
+use __PHP_Incomplete_Class;
+use RuntimeException;
 
 class Native implements Serializer
 {
@@ -21,10 +23,32 @@ class Native implements Serializer
      * Unserialize a Message.
      *
      * @param string $data
-     * @return \Rubix\Server\Message;
+     * @return \Rubix\Server\Message
      */
     public function unserialize(string $data) : Message
     {
-        return unserialize($data);
+        $message = unserialize($data);
+
+        if ($message === false) {
+            throw new RuntimeException('Cannot read encoding, wrong'
+                . ' format or corrupted data.');
+        }
+
+        if (!is_object($message)) {
+            throw new RuntimeException('Unserialized encoding must'
+                . ' be an object.');
+        }
+
+        if ($message instanceof __PHP_Incomplete_Class) {
+            throw new RuntimeException('Missing class definition'
+                . ' for unserialized object.');
+        }
+
+        if (!$message instanceof Message) {
+            throw new RuntimeException('Unserialized object must'
+                . ' be a message.');
+        }
+
+        return $message;
     }
 }